// src/pages/gestpro/components/types/ColetaSeparandoPendencias.types.ts
export type PendenciaGestproRow = {
  NUMPED: number;
  CODCLI: number;
  CLIENTE: string;
  VLTOTAL: number;
  CODFILIAL: string;
  DATA: string;
  CODUSUR: number;
  LOG2: string;
  LOG2_REAL?: string;
  TIPOENTREGA?: string;
  CODFILIALRETIRA?: string;
  POSICAO: string;
  CODPROD: number;
  QT: number;
  DESCRICAO: string;
  CODAUXILIAR: string;
  MULTIPLO?: number;
  EMBALAGEMMASTER?: string | number;
  QTD_TOTAL?: string;
  NOME?: string;
  ENDERENT?: string;
  NUMEROENT?: string;
  BAIRROENT?: string;
  MUNICENT?: string;
  CEP?: string;
  OBS?: string;
  OBS1?: string;
  OBS2?: string;
  OBSENTREGA1?: string;
  OBSENTREGA2?: string;
  OBSENTREGA3?: string;
  SEPERADOR?: string;
  NUMVIASMAPASEP?: number;
  EMISSOR_MAPA?: string;
  NUMVIASMAPASEP2?: number;
};